refactor: split monolithic cli.go into per-domain command files
Extract command struct definitions and Run() methods from the 2436-line
cli.go into 15 domain-specific files (command_databases.go,
command_tables.go, etc.). The root CLI struct, shared types (Flags,
CLIContext), helper functions, and configuration methods remain in
cli.go (~156 lines).
